aula 54 sem alert suceso no index

+++ app/Controllers/Admin/Usuarios.php
@@ -16,7 +16,7 @@ class Usuarios extends BaseController
     {
         $data = [
             'titulo' => 'Listando os usuários',
-            'usuarios' => $this->usuarioModel->findAll(),
+            'usuarios' => $this->usuarioModel->withDeleted(true)->findAll(),
         ];
        
         session()->remove('sucesso');
@@ -64,6 +64,10 @@ class Usuarios extends BaseController
 
         $usuario = $this->buscaUsuarioOu404($id);
 
+        if($usuario->is_admin){
+            return redirect()->back()->with('info', 'Não é possível excluir um <b>Administrador</b>');
+        }
+
         if($this->request->getMethod() === 'post'){
             $this->usuarioModel->delete($id);
             return redirect()->to(site_url('admin/usuarios'))->with('sucesso', "Usuário $usuario->nome excluído com sucesso.");
@@ -78,6 +82,26 @@ class Usuarios extends BaseController
         return view('Admin/Usuarios/excluir', $data);
     
     }
+
+    
+public function desfazerExclusao($id = null){
+
+    $usuario = $this->buscaUsuarioOu404($id);
+
+    if($usuario->deletado_em == null ){
+
+        return redirect()->back()->with('info', 'Apenas usuários excluídos podem ser recuperados.');
+
+    }
+
+    if($this->usuarioModel->desfazerExclusao($id)){
+        return redirect()->back()->with('sucesso', 'Exclusão desfeita com sucesso.');
+    } else {
+        return redirect()->back()->with('errors_model', $this->usuarioModel->errors())->with('atencao', 'Por favor verifique os erros abaixo.')->withInput();
+    }
+
+
+}
 
     /**
      * @uso Controller usuarios no metodo procurar com autocomplete
@@ -175,7 +199,7 @@ public function editar($id = null){
 
 
 private function buscaUsuarioOu404(int $id = null){
-    if(!$id || !$usuario = $this->usuarioModel->where('id', $id)->first()){
+    if(!$id || !$usuario = $this->usuarioModel->withDeleted(true)->where('id', $id)->first()){
         throw \CodeIgniter\Exceptions\PageNotFoundException::forPageNotFound("Não Encontramos o Usuário $id");
     }
     return $usuario;
